`/player` command group
